backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
